requirements.txt
tensorflow-cpu
opencv-python
streamlit
numpy
scikit-learn
Pillow
